requirements.txt
appnope==0.1.4
asttokens==2.4.1
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
folium==0.17.0
future==1.0.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
iso8601==2.1.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
libusb1==3.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pynmea2==1.19.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-usbtmc==0.8
pytz==2024.1
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
serial==0.0.97
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
usb==0.0.83.dev0
wcwidth==0.2.13
xyzservices==2024.6.0
